requirements.txt
qrcode>=7.3
pillow>=9.0.0
numpy>=1.20.0 